Refactor HabitSlap infrastructure: Remove Stage2Stack, enhance Stage1Stack with new Lambda function and layer, and update dependencies
- Removed Stage2Stack from app.py and the stacks directory.
- Enhanced Stage1Stack to include a new Sending Lambda function and a Lambda Layer for dependencies.
- Updated requirements.txt for both the main application and the Lambda layer with specific package versions.
- Improved .gitignore to include AWS Lambda Layer files.
- Refactored lambda_function.py to import email generation functions, streamlining the code.
- Updated generate_layer.sh to optimize the Lambda Layer creation process by removing unnecessary files.
