+++ sending_function/lambda_layer/requirements.txt
@@ -1,4 +1,5 @@
-langchain==0.1.0
-openai==1.3.0
+openai==1.59.9
 yagmail==0.15.293
-keyring==24.3.0 
+python-dotenv==1.0.0
+typing_extensions==4.12.2
+requests==2.31.0
